Extracting the `Committer` responsibility out of `StateComputer`.
